+++ tests/dlq/test_sqlite.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 
 from freezegun import freeze_time
 
-from missive import GenericMessage
+from missive import RawMessage
 from missive.dlq.sqlite import SQLiteDLQ
 
 
@@ -14,7 +14,7 @@ def dlq():
 
 @freeze_time("2018-01-03")
 def test_set_and_get(dlq):
-    message = GenericMessage(raw_data=b"test")
+    message = RawMessage(raw_data=b"test")
 
     dlq[message.message_id] = (message, "no reason")
 
@@ -27,7 +27,7 @@ def test_set_and_get(dlq):
 
 @freeze_time("2018-01-03")
 def test_set_and_remove(dlq):
-    message = GenericMessage(raw_data=b"test")
+    message = RawMessage(raw_data=b"test")
 
     dlq[message.message_id] = (message, "no reason")