+++ tests/test_matching.py
@@ -6,21 +6,19 @@ from .matchers import always, never
 
 
 def test_one_matching_handler():
-    processor: m.Processor[m.GenericMessage] = m.Processor()
+    processor: m.Processor[m.RawMessage] = m.Processor()
 
     flag = False
 
     @processor.handle_for(always)
-    def flip_bit(
-        message: m.GenericMessage, ctx: m.HandlingContext[m.GenericMessage]
-    ) -> None:
+    def flip_bit(message: m.RawMessage, ctx: m.HandlingContext[m.RawMessage]) -> None:
         nonlocal flag
         flag = True
         ctx.ack(message)
 
     test_client = processor.test_client()
 
-    blank_message = m.GenericMessage(b"")
+    blank_message = m.RawMessage(b"")
 
     test_client.send(blank_message)
 
@@ -29,7 +27,7 @@ def test_one_matching_handler():
 
 
 def test_no_matching_handler():
-    processor: m.Processor[m.GenericMessage] = m.Processor()
+    processor: m.Processor[m.RawMessage] = m.Processor()
 
     @processor.handle_for(never)
     def non_matching_handler(message, ctx):
@@ -37,7 +35,7 @@ def test_no_matching_handler():
 
     test_client = processor.test_client()
 
-    blank_message = m.GenericMessage(b"")
+    blank_message = m.RawMessage(b"")
 
     with pytest.raises(RuntimeError):
         test_client.send(blank_message)
@@ -46,7 +44,7 @@ def test_no_matching_handler():
 
 
 def test_multiple_matching_handlers():
-    processor: m.Processor[m.GenericMessage] = m.Processor()
+    processor: m.Processor[m.RawMessage] = m.Processor()
 
     @processor.handle_for(always)
     def a_matching_handler(message, ctx):
@@ -58,7 +56,7 @@ def test_multiple_matching_handlers():
 
     test_client = processor.test_client()
 
-    blank_message = m.GenericMessage(b"")
+    blank_message = m.RawMessage(b"")
 
     with pytest.raises(RuntimeError):
         test_client.send(blank_message)
@@ -67,7 +65,7 @@ def test_multiple_matching_handlers():
 
 
 def test_one_matching_handler_among_multiple():
-    processor: m.Processor[m.GenericMessage] = m.Processor()
+    processor: m.Processor[m.RawMessage] = m.Processor()
 
     flag = False
 
@@ -83,7 +81,7 @@ def test_one_matching_handler_among_multiple():
 
     test_client = processor.test_client()
 
-    blank_message = m.GenericMessage(b"")
+    blank_message = m.RawMessage(b"")
 
     test_client.send(blank_message)